CSBfilter
CSBfilter is a set of R functions written to help improve genome assembly completeness metrics. Specifically, the functions rely on BUSCO genes and sequence lengths to output a list of sequences that should be kept. This list can then be used in an external program like seqtk or samtools to subset the assembly to only those sequences listed.